commit .10

+++ software (c,c++,assembly)/001 (example of simple program in window)/code/lib/Upper0.h
@@ -34,7 +34,7 @@ class SDL_TOOLS
 {
 
 private:
-    uint64_t FpsTimerStart;
+    uint32_t FpsTimerStart;
     int32_t FpsTimerEnd;
 
 protected:
